requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
ffmpeg-python==0.2.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
PyYAML==6.0.2
requests==2.32.3
smmap==5.0.1
urllib3==2.2.3
